Add files via upload
I've kinda gotten an interest for this today, and so decided to play it. It looks pretty good, but there were some problems as it was hard to see the items (my computer couldn't display them in scale 3, even in scale 2 it was pretty low). The GUI does look a bit messy (which I am trying to somehow fix). And you couldn't respawn once you died (the engine didn't set the Scene.getInstance().getPlayer().dead to false).

Here's some of the main fixes I made:

   1. Sets the display's name to "Quake of the Chambered" by Messages.getString(), in which I added a variable that the engine finds where it sets the display name

   2. The internal yy value in Screen's function 'render' has been updated where the items are a lot more visible, but aren't where you can see the end point of the image.

   3. The yy value also has a little more exaggerated animation to better reflect the movement speed.

   4. Checks if the game is paused to allow your cursor to be unlocked from the screen and shows your cursor.

   5. Also checks if you're focused on the screen when playing the game, and which it also allows you to unlock your mouse.

   6. An add of a sprite sheet called 'guicomponents' to make up the future GUI.

   7. Fixed a respawning issue where it still considers you dead.

   8. Added fps to the frames counter as for less confusion.

If there's anything I forgot or left out, mention it.